requirements.txt
google-cloud-storage
google-cloud-speech
google-cloud-language
flask
autopep8
functions-framework
firebase-admin
mutagen
python-dotenv
flask-cors
googletrans
pandas
sendgrid
wheel